Update plexmediaserver
Without the proposed change, I found myself having to reinstall plex. Reason is that I attempted to switch the database directory after the directory was already removed. In this scenario, you cannot change the database to a new valid directory because the script crashes when it tried to remove the old directory (no longer exists)